designs/code/erms-object-detection/server/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
pillow
numpy
torch